+++ ECommerce/tests/ECommerce.UnitTests/ProductServiceTests/when_creating_a_product_with_nonexisting_category.cs
@@ -10,7 +10,7 @@ namespace ECommerce.UnitTests.ProductServiceTests
 {
     public class when_creating_a_product_with_nonexisting_category
     {
-        private decimal price = 35m;
+        private double price = 35;
         private string productTitle = "ProductTitle";
         private string categoryTitle = "CategoryTitle";
         private ProductService productService;
